+++ packages/create-stylable-app/template/ts-react-rollup/src/app.st.css
