+++ packages/create-stylable-app/template/ts-react-rollup/src/globals.st.css
