+++ packages/create-stylable-app/template/ts-react-rollup/src/header.tsx
