+++ packages/create-stylable-app/template/ts-react-rollup/src/reset.st.css
