+++ packages/create-stylable-app/template/ts-react-webpack/src/app.st.css
