+++ packages/create-stylable-app/template/ts-react-webpack/src/globals.st.css
