+++ packages/create-stylable-app/template/ts-react-webpack/src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { App } from './components/app';
-import './styles/globals.st.css';
-import './styles/reset.st.css';
+import { App } from './app';
+import './globals.st.css';
+import './reset.st.css';
 
 ReactDOM.render(
     <React.StrictMode>
