+++ packages/create-stylable-app/template/ts-react-webpack/src/reset.st.css
